xrstring_manager.h created
