+++ src/xrCore/xrstring_manager.cpp
@@ -1,0 +1,1 @@
+#include "xrstring_manager.h"
